+++ delphyne-demos/demos/mobil_perf.py
@@ -231,7 +231,7 @@ def main():
             # run for a finite amount of time
             print("Running simulation for {0} seconds.".format(args.duration))
             simulation_tree.tick_tock(
-                period=tree_time_step, number_of_iterations=args.duration / tree_time_step
+                period=tree_time_step, number_of_iterations=int(args.duration / tree_time_step)
             )
         # stop simulation if it's necessary
         if simulation_tree.runner.is_interactive_loop_running():
